feat(merge sort): add the merge sort algorithm
Here, merge sort is implemented as an recursive algorithm based on
the divide and conquer principles.
The input array is divide until the subarray contains only one element.
That element is automatically sorted, then the algorithm works inside
out by merging two adjacent subarrays. By comparing the subarrays values
they can be inserted in order in a third array.
That third array will break this iteration and move up in the recursive
order.